WIP. Refactoring and comments.
