+++ publiccode.go
@@ -5,8 +5,11 @@ import (
 	"time"
 )
 
+// Version of the PublicCode specs.
+// Source https://github.com/publiccodenet/publiccode.yml
 const Version = "0.1"
 
+// A PublicCode represents a standard metadata description for public software and policy repositories.
 type PublicCode struct {
 	Version     string
 	Url         *url.URL
@@ -37,6 +40,7 @@ type PublicCode struct {
 		Released    time.Time
 		Platforms   string
 	}
+
 	Meta struct {
 		Scope    []string
 		PaType   []string
@@ -44,6 +48,7 @@ type PublicCode struct {
 		Tags     []string
 		UsedBy   []string
 	}
+
 	Dependencies struct {
 		Hardware    []string
 		Open        []string
@@ -51,12 +56,14 @@ type PublicCode struct {
 	}
 }
 
+// A Contact represents all the standard informations about a technical contact.
 type Contact struct {
 	Name        string
 	Email       string
 	Affiliation string
 }
 
+// A Desc represents a generic description with multiple languages.
 type Desc struct {
 	En string
 	It string
